Move search service into services folder
