+++ webapp/src/app/services/search-filter.service.ts
