+++ webapp/src/app/shared/list-search-overlay/list-search-overlay.component.ts
@@ -1,6 +1,6 @@
 import { Component, ElementRef, EventEmitter, Input, OnInit, Output, ViewChild } from '@angular/core';
 import { animate, state, style, transition, trigger } from '@angular/animations';
-import { SearchFilterService } from '../search-filter.service';
+import { SearchFilterService } from '../../services/search-filter.service';
 
 const ANIMATION_TIMING = '300ms cubic-bezier(0.25, 0.8, 0.25, 1)';
